### 2024-11-12 #7456 @editorialbot
[PRE REVIEW]: ERAD: A Graph-Based Tool for Equity and Resilience Analysis of electric Distribution systems
**Submitting author:** @KapilDuwadi (<a href="http://orcid.org/0000-0002-0589-5187">KAPIL DUWADI</a>)
**Repository:** https://github.com/NREL/erad
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
