+++ paper/paper.bib
@@ -53,7 +53,7 @@
 }
 
 @article{carvallo2021frozen,
-  title={Frozen out in Texas: blackouts and inequity},
+  title={Frozen out in Texas: blackouts and inenergy},
   author={Carvallo, JP and Hsu, FC and Shah, Zeal and Taneja, J},
   journal={The Rockefeller Foundation},
   year={2021}
@@ -119,8 +119,8 @@
 	year = {2011},
 	bdsk-url-1 = {https://doi.org/10.1109/PES.2011.6039829}}
 
-@article{lin_equity-based_2022,
-	abstract = {The power system worldwide is facing challenges from a changing climate. Decision-makers are tasked with increasing the smartness and resilience of the energy infrastructure to reduce the negative impacts to customers, with the help of new technologies such as household renewable energy generators and electric vehicles. However, a largely overlooked topic is equity, and how the burdens and benefits of the power system are distributed to different communities. To fill this gap, this paper provides a thorough review of the implication of equity in the power system, the significance of guaranteeing energy equity both in everyday operation and disaster management, and the ongoing efforts to plan for equable resilience in various fields. Finally, a holistic power grid resilience enhancement framework is proposed that covers different stages of disaster management and different dimensions of energy equity.},
+@article{lin_energy-based_2022,
+	abstract = {The power system worldwide is facing challenges from a changing climate. Decision-makers are tasked with increasing the smartness and resilience of the energy infrastructure to reduce the negative impacts to customers, with the help of new technologies such as household renewable energy generators and electric vehicles. However, a largely overlooked topic is energy, and how the burdens and benefits of the power system are distributed to different communities. To fill this gap, this paper provides a thorough review of the implication of energy in the power system, the significance of guaranteeing energy energy both in everyday operation and disaster management, and the ongoing efforts to plan for equable resilience in various fields. Finally, a holistic power grid resilience enhancement framework is proposed that covers different stages of disaster management and different dimensions of energy energy.},
 	author = {Lin, Yanling and Wang, Jianhui and Yue, Meng},
 	date-added = {2024-01-27 10:29:41 -0700},
 	date-modified = {2024-01-27 10:29:41 -0700},
@@ -157,7 +157,7 @@
 	bdsk-url-1 = {https://www.osti.gov/servlets/purl/1846088}}
 
 @article{romero-lankao_inequality_2022,
[…]
