+++ src/erad/scenarios/flood_scenario.py
@@ -1,0 +1,419 @@
+import tarfile
+import os
+
+from shapely import MultiPolygon, Point, LineString
+from pyhigh import get_elevation, get_elevation_batch
+from datetime import datetime, timedelta
+from scipy.spatial import Delaunay
+import matplotlib.pyplot as plt
+import geopandas as gpd
+import pandas as pd
+import numpy as np
+import stateplane
+import itertools
+import requests
+
+from erad.scenarios.utilities import ProbabilityFunctionBuilder, GeoUtilities
+from erad.scenarios.utilities import ProbabilityFunctionBuilder
+from erad.constants import DATA_FOLDER, FLOOD_HISTORIC_SHP_PATH
+from erad.scenarios.abstract_scenario import BaseScenario
+from erad.scenarios.common import AssetTypes
+
+plt.ion()
+class FloodScenario(BaseScenario, GeoUtilities):
+    """Base class for FlooadScenario. Extends BaseScenario and GeoUtilities
+
+    Attributes:
+        origin (Point): Earthquake origin point
+        probability_model (dict): Dictionary mapping asset types to probability funcitons
+        timestamp (datetime): Scenario occurance time 
+        kwargs (dict): Additional parameters relevant for a particular scenario type      
+    """
+    
+    fragility_curves = {
+        # Electrical Grid Risk Assessment Against Flooding in Barcelona and Bristol Cities
+        AssetTypes.substation.name : ProbabilityFunctionBuilder("norm", [1.5, 0.2]),
+        AssetTypes.solar_panels.name : ProbabilityFunctionBuilder("norm", [0.04, .01]),
+        # AssetTypes.buried_lines.name : ProbabilityFunctionBuilder("norm", [80, 10]),
+        # estimated
+        AssetTypes.wind_turbines.name : ProbabilityFunctionBuilder("lognorm", [0.8, 3, 2]),
+        AssetTypes.battery_storage.name : ProbabilityFunctionBuilder("norm", [0.04, .01]),
+        # estimated
+        AssetTypes.transmission_poles.name  :  ProbabilityFunctionBuilder("lognorm", [0.8, 5, 3]),
+        # Tsunami Fragility Functions for Road and Utility Pole Assets Using Field Survey and Remotely Sensed Data from the 2018 Sulawesi Tsunami, Palu, Indonesia
+        AssetTypes.distribution_poles.name  : ProbabilityFunctionBuilder("lognorm", [1, 0.2, 1]),
+        AssetTypes.transmission_overhead_lines.name  : ProbabilityFunctionBuilder("norm", [0.04, .01]),    
+        AssetTypes.distribution_overhead_lines.name  : ProbabilityFunctionBuilder("norm", [0.04, .01]),
+    }
+    
+    
+    def __init__(self,  poly : MultiPolygon , probability_model : dict, timestamp : datetime, **kwargs) -> None:
+        super(FloodScenario, self).__init__(poly, probability_model, timestamp, **kwargs)
+        self.kwargs = kwargs
+        self.samples = 20
+    
+        if 'type' in kwargs and kwargs['type'] == 'live':
+            self.flows = pd.DataFrame()
+            self.levels = pd.DataFrame()
+            self.real_time()
+        else:
+            from shapely import wkt
+            self.flows = pd.read_csv(kwargs['file_flow'],index_col=0, parse_dates=True)
+            self.levels = pd.read_csv(kwargs['file_levels'], index_col=0, parse_dates=True)
+            df = pd.read_csv(kwargs['file_gaugues'])
+            df['geometry'] = df['geometry'].apply(wkt.loads)
+            self.gauges = gpd.GeoDataFrame(df).set_geometry('geometry')
+
[…]
