+++ db_neo4j/db_neo4j.md
@@ -1,1 +1,0 @@
-::: erad.db.neo4j_
