+++ exceptions/exceptions.md
@@ -1,1 +1,0 @@
-::: erad.exceptions
