+++ utils_hifld_utils/utils_hifld_utils.md
@@ -1,1 +1,0 @@
-::: erad.utils.hifld_utils
