+++ utils_opendss_utils/utils_opendss_utils.md
@@ -1,1 +1,0 @@
-::: erad.utils.opendss_utils
