+++ utils_ditto_utils/utils_ditto_utils.md
@@ -1,1 +1,0 @@
-::: erad.utils.ditto_utils
