+++ paper/paper.bib
@@ -167,7 +167,7 @@
 	file = {ScienceDirect Snapshot:/Users/bpalmint/Zotero/storage/PPEKU27W/S2214629622002638.html:text/html},
 	issn = {2214-6296},
 	journal = {Energy Research \& Social Science},
-	keywords = {Health risks, Cities, Electric vehicle adoption, Equitable and just transitions, Transportation energy, Typologies},
+	keywords = {Health risks, Cities, Electric vehicle adoption, Energy and just transitions, Transportation energy, Typologies},
 	language = {en},
 	month = sep,
 	pages = {102760},
@@ -353,7 +353,7 @@ These results may assist public health jurisdictions addressing unique needs and
 	month = oct,
 	note = {NREL/TP-6A40-85959},
 	publisher = {National Renewable Energy Lab},
-	title = {Chapter 12. {Distribution} {Grid} {Upgrades} for {Equitable} {Resilience} and {Solar}, {Storage}, and {Electric} {Vehicle} {Access}},
+	title = {Chapter 12. {Distribution} {Grid} {Upgrades} for {Energy} {Resilience} and {Solar}, {Storage}, and {Electric} {Vehicle} {Access}},
 	year = {2023}}
 
 @techreport{la100-es-report,
