+++ pyproject.toml
@@ -21,7 +21,7 @@ maintainers = [
 keywords = [
     "Distribution",
     "Earthquake",
-    "Equity",
+    "Energy",
     "Fire",
     "Flooding",
     "Power",
