### 2025-04-14 #7456 @anastassiavybornova
Editor checklist (@anastassiavybornova)

**Software**
- [ ] ∃ software description
- [ ] ∃ OSI-approved license
- [ ] ∃ installation instructions
- [ ] ∃ documentation
- [ ] ∃ tests

- [ ] Paper compiles
- [ ] Paper length 250-1000 words

**Application**
- [ ] Substantial scholarly effort
