### 2025-06-27 #7456 @aslan-ng
Dear Anastassia (@anastassiavybornova),

I have relevant experience in modeling network resilience, and I’d be happy to serve as a reviewer for this submission if you approve.

Best regards,
Aslan (@aslan-ng)
