### 2025-07-18 #7456 @AadilLatif
@anastassiavybornova, the new version with documentation, tests and paper updates. We are ready for review. (@KapilDuwadi)
