### 2025-07-19 #7456 @AadilLatif
**the word count does not exceed 1000 words (currently 1290)**
_Reduced to less than 1000 works_

**the link to full documentation under https://github.com/NREL-Distribution-Suites/erad leads to the actual full documentation (it appears broken to me)**
_New documentation has been uploaded_

**the references are in correct format, meaning:**

> **check the references signalled as potentially erroneous by the editorialbot (see comment above this one in this issue)**
> I have added DOI to most referenced. A book had an ISBN but not a DOI. For a couple of references I was not able to find te DOI. Please let me know how to procede
> **instead of name, (name, year) it should be: name, year everywhere (please review the commands with which you cite, and don't use parentheses when authors are directly cited by name).**
> _fixed_
> **For example instead of: "Liu, et al (Liu et al., 2017) say..." in the text it should be: "Liu et al. (2017) say..." where "Liu et al. (2017)" doesn't need to be typed out by you but can be generated with the citation commands you're using**
> _fixed_
>
> **check the references appearing as [?] in the PDF**
> _fixed_
>
> **correct the broken markdown link: ([HIFLD])(https://hifld-geoplatform.opendata.arcgis.com/)**
> _removed_
>
> **when citing several references, they all go into the same parentheses group (ref1, ref2, ...) instead of (ref1), (ref2)**
> _Fixed_
