### 2025-07-24 #7456 @anastassiavybornova
hi @AadilLatif , paper length and references look good now, and i managed to find the full documentation here: https://nrel-distribution-suites.github.io/erad/intro.html , will start looking for reviewers now, please in the meantime fix the broken link `"[Visit full documentation here.](https://nrel.github.io/erad/)"` on the README page here: https://github.com/NREL-Distribution-Suites/erad
