### 2025-07-29 #7456 @AAndersn
Hi @anastassiavybornova , I'll have to decline this review due to conflict of interest. When this paper was submitted, Kapil was at NREL, but now he is in the same division as me at PNNL. However, I'd be interested in contributing as a reviewer to other papers in this domain.
