### 2025-07-29 #7456 @anastassiavybornova
@AAndersn huge thanks! and appreciate your pointing out the CoI. in case any non-conflicted colleague comes to mind, you're more than welcome to share their names here (without ping'ing them yet), that would also be of great help for us to find reviewers. and yes, definitely we'll be in touch from JOSS side for future potential reviews!
