### 2025-08-12 #8782 @editorialbot
[REVIEW]: ERAD: A Graph-Based Tool for Equity and Resilience Analysis of electric Distribution systems
**Submitting author:** @KapilDuwadi (<a href="http://orcid.org/0000-0002-0589-5187">KAPIL DUWADI</a>)
**Repository:** https://github.com/NREL/erad
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.7
