### 2025-08-16 #8782 @aslan-ng
Hi Aadil (@AadilLatif) and the rest of the authors,

Great work on ERAD! The tool addresses an important challenge in analyzing and enhancing community-level energy resilience. I also found the codebase neat, well-structured, and easy to follow. It’s clear that a lot of thought and craft went into the development.

I’ve organized my review comments into three sections for easier navigation:
- Readme and Docs
- Code
- Paper

Thanks again for developing and open-sourcing ERAD. It’s already a valuable toolkit, and I’m excited to see it gain adoption in the energy resilience community.

Best,
Aslan (@aslan-ng)
