### 2025-08-16 #8782 @aslan-ng
# Code

5. I recommend adding a few concrete usage examples, ideally collected in an examples/ folder in the repository. They could walk through simple, real-world scenarios and demonstrate how to translate them into ERAD code. Such examples would greatly help new users understand how to apply the package to their own workflows. The examples don’t need to be complex, since more advanced research applications are already described in the papers that used this software package for their analysis. Rather, the goal would be to provide quick, accessible entry points for first-time users.

6. Since ERAD is based on a graph model, would it be possible to expose or export the results in the form of network data? This could enable users to extend the analysis beyond what is currently built into the package and integrate ERAD outputs with other graph/network analysis tools.

7. Python version metadata mismatch:
In `pyproject.toml`, you declare:
```
requires-python = ">=3.11"
```
But then list this classifier:
```
"Programming Language :: Python :: 3.8"
```
This inconsistency may confuse users about which Python versions are actually supported. I recommend either removing the 3.8 classifier altogether or updating your classifiers to match the declared requirement. This will ensure that your package metadata accurately reflects the versions you guarantee support for.
Additionally, I tested installation with Python 3.13 and it does not appear to run, while 3.11 works as expected. It may be helpful to specify the supported upper bound (e.g., <3.13) if you have not yet validated compatibility with newer versions.
