### 2025-08-16 #8782 @aslan-ng
# Paper

8. Since your project has a polished and professional logo, you can include it as a figure in your paper if you wish:

![logo](https://github.com/NREL-Distribution-Suites/erad/blob/main/docs/light.png?raw=true)

For inspiration, see how other JOSS papers have showcased theirs, for example:

* [LightLogR: Reproducible analysis of personal light exposure data](https://joss.theoj.org/papers/10.21105/joss.07601)
* [madupite: A High-Performance Distributed Solver for Large-Scale Markov Decision Processes](https://joss.theoj.org/papers/10.21105/joss.07411)
* [STUMPY: A Powerful and Scalable Python Library for Time Series Data Mining](https://joss.theoj.org/papers/10.21105/joss.01504)
* [polyCub: An R package for Integration over Polygons](https://joss.theoj.org/papers/10.21105/joss.01056)

9. There is a typo in the keywords list: “resilince” should be corrected to “resilience.”

10. A typo:
“induvisual” -> “individual”

11. I’d suggest updating the title for consistent capitalization and dropping the trailing period.
ERAD: A Graph-Based Tool for Energy Resilience Analysis of <u>e</u>lectric Distribution <u>s</u>ystems<u>.</u>' -> 'ERAD: A Graph-Based Tool for Energy Resilience Analysis of Electric Distribution Systems'
