### 2025-09-01 #8782 @KathiEsterl
Hi @AadilLatif and the other authors,

First, thank you very much for your contribution – the topic you are addressing is both exciting and highly relevant. I found the code overall well-structured and understandable, and the resulting publications are interesting to read.

I would like to echo some of the suggestions already raised by @aslan-ng. In particular:

Providing **example scripts** would significantly lower the entry barrier for new users. Ideally, one example could demonstrate usage with a GDM model and another without, with clear references to these examples in the documentation.

The **documentation** could be made even more helpful by being structured more clearly. For instance:

- A table of contents on the start page with subchapters.
- Currently, the flow goes from the hazard model directly to the simulation. It might be more intuitive to first present the asset model (and possible extensions with the GDM model) on the modeling page, and then continue with calculations and analyses. This would keep the content more cleanly separated, I think.
- The GDM connection could perhaps already be introduced at the installation stage.

Regarding the **paper**, I believe there is still room to make it more compelling:

- Ensure consistent linking of citations.
- Avoid clustering multiple references to a single statement, as this makes attribution harder to follow.
- The writing style could be smoothed out in places. For example, in the "Implementation Overview," the second paragraph reads more like a newly added fragment than as part of a coherent narrative.

In general, I found that **more explicit references to sources** would strengthen both the documentation and the paper. For example, when trying out the tool for the first time using the code provided in the documentation, some exemplary data was downloaded. It would be very helpful to clearly indicate the origin of this data – both for the hazard data as well as the asset system data. If there is a key publication that addresses these aspects, it should be cited explicitly to guide readers and users more effectively.

Overall, I think these adjustments would further enhance the accessibility and impact of your work. I hope this feedback is helpful, and I look forward to your response!
