### 2025-09-04 #8782 @AadilLatif
> 4\. he “CodeFactor” badge in the README points to a broken URL. Could you update the badge target so it matches the current org/repo path, or remove the badge if the project isn’t tracked on CodeFactor anymore?

All suggested changes have been made
