### 2025-10-21 #8782 @KathiEsterl
Dear Anastassia (@anastassiavybornova ),

I have re-evaluated the paper and repository and confirm that all comments have been fully addressed. The authors have responded thoughtfully and improved the work. I therefore recommend this submission for publication in JOSS.

Kind regards,
Kathi
