### 2025-11-26 #8782 @anastassiavybornova
@AadilLatif great! last thing from my side: the author order at the zenodo archive has to match the one on the paper (since one is the archive of the other), please update the zenodo archive author order and ping me here once you're done! 🙏
