### 2025-11-27 #8782 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
