### 2025-12-04 #8782 @kyleniemeyer
@KapilDuwadi my only comment on the paper before accepting is that the Alper et al. reference seems to be missing information (journal name, volume, pages, etc.) Can you fix this?
